Move the relocate module from the rule folder to the ir folder
