+++ ir/rule/conditional/if.go
@@ -19,8 +19,8 @@ import (
 	"fluent/filecode"
 	"fluent/filecode/types/wrapper"
 	"fluent/ir/pool"
+	"fluent/ir/relocate"
 	"fluent/ir/rule/expression"
-	"fluent/ir/rule/relocate"
 	"fluent/ir/tree"
 	"fluent/ir/value"
 	"fluent/util"
