+++ ir/rule/loop/while.go
@@ -18,9 +18,9 @@ import (
 	"fluent/ast"
 	"fluent/filecode"
 	"fluent/ir/pool"
+	"fluent/ir/relocate"
 	"fluent/ir/rule/conditional"
 	"fluent/ir/rule/expression"
-	"fluent/ir/rule/relocate"
 	"fluent/ir/tree"
 	"fluent/ir/value"
 	"fluent/util"
